*** empty log message ***

+++ source/ch/cyberduck/core/sftp/SFTPPath.java
@@ -186,6 +186,7 @@ public class SFTPPath extends Path {
 			session.check();
 			session.log("Make directory "+this.getName(), Message.PROGRESS);
 			session.SFTP.makeDirectory(this.getAbsolute());
+			session.cache().put(this.getAbsolute(), new ArrayList());
 			this.getParent().invalidate();
 			session.log("Idle", Message.STOP);
 		}
@@ -288,38 +289,43 @@ public class SFTPPath extends Path {
 		OutputStream out = null;
 		try {
 			session.check();
-			out = new FileOutputStream(this.getLocal().getTemp(), this.status.isResume());
-			if(out == null) {
-				throw new IOException("Unable to buffer data");
-			}
-			SftpFile p = session.SFTP.openFile(this.getAbsolute(), SftpSubsystemClient.OPEN_READ);
-			in = new SftpFileInputStream(p);
-			if(in == null) {
-				throw new IOException("Unable opening data stream");
-			}
-			if(this.status.isResume()) {
-				this.status.setCurrent(this.getLocal().getTemp().length());
-				long skipped = in.skip(this.status.getCurrent());
-				log.info("Skipping "+skipped+" bytes");
-				if(skipped < this.status.getCurrent()) {
-					throw new IOException("Resume failed: Skipped "+skipped+" bytes instead of "+this.status.getCurrent());
+			if(this.attributes.isFile()) {
+				out = new FileOutputStream(this.getLocal().getTemp(), this.status.isResume());
+				if(out == null) {
+					throw new IOException("Unable to buffer data");
 				}
-			}
-			this.download(in, out);
-			if(this.status.isComplete()) {
-				log.info("Updating permissions");
-				if(Preferences.instance().getProperty("queue.download.changePermissions").equals("true")) {
-					Permission perm = null;
-					if(Preferences.instance().getProperty("queue.download.permissions.useDefault").equals("true")) {
-						perm = new Permission(Preferences.instance().getProperty("queue.download.permissions.default"));
-					}
-					else {
-						perm = this.attributes.getPermission();
-					}
-					if(!perm.isUndefined()) {
-						this.getLocal().setPermission(perm);
+				SftpFile p = session.SFTP.openFile(this.getAbsolute(), SftpSubsystemClient.OPEN_READ);
+				in = new SftpFileInputStream(p);
+				if(in == null) {
+					throw new IOException("Unable opening data stream");
+				}
+				if(this.status.isResume()) {
+					this.status.setCurrent(this.getLocal().getTemp().length());
+					long skipped = in.skip(this.status.getCurrent());
+					log.info("Skipping "+skipped+" bytes");
+					if(skipped < this.status.getCurrent()) {
+						throw new IOException("Resume failed: Skipped "+skipped+" bytes instead of "+this.status.getCurrent());
 					}
 				}
+				this.download(in, out);
+				if(this.status.isComplete()) {
+					log.info("Updating permissions");
+					if(Preferences.instance().getProperty("queue.download.changePermissions").equals("true")) {
+						Permission perm = null;
+						if(Preferences.instance().getProperty("queue.download.permissions.useDefault").equals("true")) {
+							perm = new Permission(Preferences.instance().getProperty("queue.download.permissions.default"));
+						}
+						else {
+							perm = this.attributes.getPermission();
+						}
+						if(!perm.isUndefined()) {
+							this.getLocal().setPermission(perm);
+						}
+					}
+				}
+			}
+			if(this.attributes.isDirectory()) {
+				this.getLocal().mkdir();
 			}
 			if(Preferences.instance().getProperty("queue.download.preserveDate").equals("true")) {
 				this.getLocal().setLastModified(this.attributes.getTimestamp().getTime());
@@ -349,55 +355,60 @@ public class SFTPPath extends Path {
 			}
 		}
 	}
-
+	
 	public synchronized void upload() {
 		log.debug("upload:"+this.toString());
 		InputStream in = null;
 		SftpFileOutputStream out = null;
 		try {
 			session.check();
-			in = new FileInputStream(this.getLocal());
-			if(in == null) {
-				throw new IOException("Unable to buffer data");
-			}
-			SftpFile p = null;
-			if(this.status.isResume()) {
-				p = session.SFTP.openFile(this.getAbsolute(),
-										  SftpSubsystemClient.OPEN_WRITE | //File open flag, opens the file for writing.
-										  SftpSubsystemClient.OPEN_APPEND); //File open flag, forces all writes to append data at the end of the file.
-			}
-			else {
-				p = session.SFTP.openFile(this.getAbsolute(),
-										  SftpSubsystemClient.OPEN_CREATE | //File open flag, if specified a new file will be created if one does not already exist.
-										  SftpSubsystemClient.OPEN_WRITE | //File open flag, opens the file for writing.
-										  SftpSubsystemClient.OPEN_TRUNCATE); //File open flag, forces an existing file with the same name to be truncated to zero length when creating a file by specifying OPEN_CREATE.
-			}
-			if(this.status.isResume()) {
-				this.status.setCurrent(p.getAttributes().getSize().intValue());
-			}
-			out = new SftpFileOutputStream(p);
-			if(out == null) {
-				throw new IOException("Unable opening data stream");
-			}
-			if(this.status.isResume()) {
-				long skipped = out.skip(this.status.getCurrent());
-				log.info("Skipping "+skipped+" bytes");
-				if(skipped < this.status.getCurrent()) {
-					throw new IOException("Resume failed: Skipped "+skipped+" bytes instead of "+this.status.getCurrent());
+			if(this.attributes.isFile()) {
+				in = new FileInputStream(this.getLocal());
+				if(in == null) {
+					throw new IOException("Unable to buffer data");
 				}
-			}
-			this.upload(out, in);
-			if(Preferences.instance().getProperty("queue.upload.changePermissions").equals("true")) {
-				Permission perm = null;
-				if(Preferences.instance().getProperty("queue.upload.permissions.useDefault").equals("true")) {
-					perm = new Permission(Preferences.instance().getProperty("queue.upload.permissions.default"));
+				SftpFile p = null;
+				if(this.status.isResume()) {
+					p = session.SFTP.openFile(this.getAbsolute(),
+											  SftpSubsystemClient.OPEN_WRITE | //File open flag, opens the file for writing.
+											  SftpSubsystemClient.OPEN_APPEND); //File open flag, forces all writes to append data at the end of the file.
 				}
 				else {
-					perm = this.getLocal().getPermission();
+					p = session.SFTP.openFile(this.getAbsolute(),
+											  SftpSubsystemClient.OPEN_CREATE | //File open flag, if specified a new file will be created if one does not already exist.
+											  SftpSubsystemClient.OPEN_WRITE | //File open flag, opens the file for writing.
+											  SftpSubsystemClient.OPEN_TRUNCATE); //File open flag, forces an existing file with the same name to be truncated to zero length when creating a file by specifying OPEN_CREATE.
 				}
-				if(!perm.isUndefined()) {
-					this.changePermissions(perm, false);
+				if(this.status.isResume()) {
+					this.status.setCurrent(p.getAttributes().getSize().intValue());
 				}
+				out = new SftpFileOutputStream(p);
+				if(out == null) {
+					throw new IOException("Unable opening data stream");
+				}
+				if(this.status.isResume()) {
+					long skipped = out.skip(this.status.getCurrent());
+					log.info("Skipping "+skipped+" bytes");
+					if(skipped < this.status.getCurrent()) {
+						throw new IOException("Resume failed: Skipped "+skipped+" bytes instead of "+this.status.getCurrent());
+					}
+				}
+				this.upload(out, in);
+				if(Preferences.instance().getProperty("queue.upload.changePermissions").equals("true")) {
+					Permission perm = null;
+					if(Preferences.instance().getProperty("queue.upload.permissions.useDefault").equals("true")) {
+						perm = new Permission(Preferences.instance().getProperty("queue.upload.permissions.default"));
+					}
+					else {
+						perm = this.getLocal().getPermission();
+					}
+					if(!perm.isUndefined()) {
+						this.changePermissions(perm, false);
+					}
+				}
+			}
+			if(this.attributes.isDirectory()) {
+				this.mkdir();
 			}
 			session.log("Idle", Message.STOP);
 		}
